requirements.txt
asgiref==3.7.2
attrs==23.1.0
coverage==7.2.7
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
flake8==6.0.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Markdown==3.4.3
mccabe==0.7.0
packaging==23.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
uritemplate==4.1.1
